Various changes (holy macaroni I think I fixed it)

+++ MapParse/MapParse/Types/Vertex.cs
@@ -6,12 +6,12 @@ namespace MapParse.Types
 	public class Vertex
 	{
 		public Vec3 P { get; set; }
-		public double[] Tex { get; set; }
+		public float[] Tex { get; set; }
 	
 		public Vertex(Vec3 v)
 		{
 			P = v;
-			Tex = new double[2];
+			Tex = new float[2];
 		}
 	
 		public Vertex(Vertex v)
@@ -23,7 +23,7 @@ namespace MapParse.Types
 		public Vertex()
 		{
 			P = new Vec3();
-			Tex = new double[2];
+			Tex = new float[2];
 		}
 	}
 }
